exclude some files domtom of process

+++ data_processing/meteo/pg_processing/task_functions.py
@@ -149,6 +149,20 @@ def download_data(ti, dataset_name):
         dates=dates,
     )
 
+def excluded_files(csv_path):
+    is_excluded = False
+    excluded_files = [
+        "MN_974_2000-2009.csv.gz",
+        "MN_974_2010-2019.csv.gz",
+        "MN_988_2000-2009.csv.gz",
+        "MN_988_2010-2019.csv.gz"
+    ]
+    for item in excluded_files:
+        if item in csv_path:
+            print("exluded - ", item)
+            is_excluded = True
+    return is_excluded
+
 
 def fetch_resources(dataset):
     r = requests.get(
@@ -220,38 +234,39 @@ def process_resources(
         _conn = psycopg2.connect(**db_params)
         _conn.autocommit = False
         try:
-            diff = get_diff(
-                _conn=_conn,
-                csv_path=csv_path,
-                regex_infos=regex_infos,
-                table=table_name,
-            )
-            delete_and_insert_into_pg(
-                _conn=_conn,
-                diff=diff,
-                regex_infos=regex_infos,
-                table=table_name,
-                csv_path=csv_path,
-            )
-            _conn.commit()
+            if not excluded_files(csv_path):
+                diff = get_diff(
+                    _conn=_conn,
+                    csv_path=csv_path,
+                    regex_infos=regex_infos,
+                    table=table_name,
+                )
+                delete_and_insert_into_pg(
+                    _conn=_conn,
+                    diff=diff,
+                    regex_infos=regex_infos,
+                    table=table_name,
+                    csv_path=csv_path,
+                )
+                _conn.commit()
 
-            minio_meteo.send_files(
-                list_files=[
-                    {
-                        "source_path": "/".join(csv_path.split("/")[:-1]),
-                        "source_name": csv_path.split("/")[-1],
-                        "dest_path": "synchro_pg/" + "/".join(resource["url"].split("data/synchro_ftp/")[1].split("/")[:-1]) + "/",
-                        "dest_name": resource["url"].split("/")[-1].replace(".csv.gz", ".csv")
-                    }
-                ],
-                ignore_airflow_env=True
-            )
+                minio_meteo.send_files(
+                    list_files=[
+                        {
+                            "source_path": "/".join(csv_path.split("/")[:-1]),
+                            "source_name": csv_path.split("/")[-1],
+                            "dest_path": "synchro_pg/" + "/".join(resource["url"].split("data/synchro_ftp/")[1].split("/")[:-1]) + "/",
+                            "dest_name": resource["url"].split("/")[-1].replace(".csv.gz", ".csv")
+                        }
+                    ],
+                    ignore_airflow_env=True
+                )
 
-            # deleting if everything was successful, so that we can check content otherwise
-            parent = file_path.parent.as_posix()
-            for file in os.listdir(parent):
-                os.remove(f"{parent}/{file}")
-            print("=> Completed work for:", regex_infos["name"])
+                # deleting if everything was successful, so that we can check content otherwise
+                parent = file_path.parent.as_posix()
+                for file in os.listdir(parent):
+                    os.remove(f"{parent}/{file}")
+                print("=> Completed work for:", regex_infos["name"])
         except Exception as e:
             _conn.rollback()
             print(f"An error occurred: {e}")
@@ -318,82 +333,32 @@ def unzip_csv_gz(file_path):
 
 def get_diff(_conn, csv_path: Path, regex_infos: dict, table: str):
 
-    def run_diff_dom_tom(csv_path: str, regex_infos):
-        print("run_diff_domtom")
-        old_file_path = csv_path.replace(".csv", "_old.csv")
-        additions_file_path = csv_path.replace(".csv", "_additions.csv")
-        deletions_file_path = csv_path.replace(".csv", "_deletions.csv")
+    def run_diff(csv_path: str, regex_infos):
+        old_file = csv_path.replace(".csv", "_old.csv")
+        additions_file = csv_path.replace(".csv", "_additions.csv")
+        deletions_file = csv_path.replace(".csv", "_deletions.csv")
         is_additions = False
 
-        with open(csv_path, 'r') as new_file, open(old_file_path, 'r') as old_file:
-            new_header = new_file.readline().strip()
+        with open(csv_path, 'r') as new_file, open(old_file, 'r') as old_file:
+            new_header = new_file.readline()
+            old_header = old_file.readline()
 
-            with open(additions_file_path, 'w') as additions_out:
-                additions_out.write(new_header + ";DEP\n")
+            new_lines = set(new_file)
+            old_lines = set(old_file)
 
-                old_line = old_file.readline().strip()
-                for new_line in new_file:
-                    new_line = new_line.strip()
+        with open(additions_file, 'w') as outFile:
+            outFile.write(new_header.strip() + ";DEP\n")
+            for line in new_lines:
+                if line not in old_lines:
+                    is_additions = True
+                    outFile.write(line.strip() + ";" + regex_infos["regex_infos"]["DEP"] + "\n")
+                
 
-                    while old_line and old_line < new_line:
-                        old_line = old_file.readline().strip()
-
-                    if old_line != new_line:
-                        is_additions = True
-                        additions_out.write(new_line + ";" + regex_infos["regex_infos"]["DEP"] + "\n")
-
-                    if old_line == new_line:
-                        old_line = old_file.readline().strip()
-
-        with open(csv_path, 'r') as new_file, open(old_file_path, 'r') as old_file:
-            new_header = new_file.readline().strip() 
-
-            with open(deletions_file_path, 'w') as deletions_out:
-                deletions_out.write(new_header)
-
-                new_lines = set(line.strip() for line in new_file)
-
-                for old_line in old_file:
-                    old_line = old_line.strip()
-                    if old_line not in new_lines:
-                        deletions_out.write(old_line + "\n")
-
-        return is_additions
-
-
-    def run_diff(csv_path: str, regex_infos):
-        if regex_infos["regex_infos"]["DEP"] not in [
-            '971', '972', '973', '974', '975',
-            '984', '985', '986', '987', '988',
-            '99'
-        ]:
-            old_file = csv_path.replace(".csv", "_old.csv")
-            additions_file = csv_path.replace(".csv", "_additions.csv")
-            deletions_file = csv_path.replace(".csv", "_deletions.csv")
-            is_additions = False
-
-            with open(csv_path, 'r') as new_file, open(old_file, 'r') as old_file:
-                new_header = new_file.readline()
-                old_header = old_file.readline()
-
-                new_lines = set(new_file)
-                old_lines = set(old_file)
-
-            with open(additions_file, 'w') as outFile:
-                outFile.write(new_header.strip() + ";DEP\n")
-                for line in new_lines:
-                    if line not in old_lines:
-                        is_additions = True
-                        outFile.write(line.strip() + ";" + regex_infos["regex_infos"]["DEP"] + "\n")
-                    
-
-            with open(deletions_file, 'w') as outFile:
-                outFile.write(old_header)
-                for line in old_lines:
-                    if line not in new_lines:
-                        outFile.write(line)
-        else:
-            is_additions = run_diff_dom_tom(csv_path=csv_path, regex_infos=regex_infos)
+        with open(deletions_file, 'w') as outFile:
+            outFile.write(old_header)
+            for line in old_lines:
+                if line not in new_lines:
+                    outFile.write(line)
         return is_additions
